+++ src/main/java/uk/q3c/krail/core/guice/DefaultBindingManager.java
@@ -32,7 +32,7 @@ import uk.q3c.krail.core.eventbus.EventBusModule;
 import uk.q3c.krail.core.guice.threadscope.ThreadScopeModule;
 import uk.q3c.krail.core.guice.uiscope.UIScopeModule;
 import uk.q3c.krail.core.guice.vsscope.VaadinSessionScopeModule;
-import uk.q3c.krail.core.i18n.VaadinI18NModule;
+import uk.q3c.krail.core.i18n.KrailI18NModule;
 import uk.q3c.krail.core.navigate.NavigationModule;
 import uk.q3c.krail.core.navigate.sitemap.MasterSitemap;
 import uk.q3c.krail.core.navigate.sitemap.SitemapModule;
@@ -266,7 +266,7 @@ public abstract class DefaultBindingManager extends GuiceServletContextListener 
      * @return a Module for I18N
      */
     protected Module i18NModule() {
-        return new VaadinI18NModule();
+        return new KrailI18NModule();
     }
 
     /**